Re-organize the components
