+++ packages/ui-core/src/components/data-access/Users.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import gql from 'graphql-tag';
 import PropTypes from 'prop-types';
 import { Query } from 'react-apollo';
-import { Loading, ErrorInfo, Page } from '../../ui-components';
+import { Loading, ErrorInfo, Page } from '../core';
 import { ResolvedUser, QueryState } from '../../helpers';
 import UsersTable from './UsersTable';
