+++ packages/ui-core/src/components/framework/index.ts
@@ -1,0 +1,11 @@
+import AppWrapper from './AppWrapper';
+import Authenticate from './Authenticate';
+import Login from './Login';
+import Welcome from './Welcome';
+
+export {
+    AppWrapper,
+    Authenticate,
+    Login,
+    Welcome,
+};
